requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
chromadb>=0.4.22
openai>=1.14.0
python-dotenv>=1.0.0
tiktoken>=0.5.2
pypdf>=3.17.1
unstructured>=0.10.30
pdf2image>=1.16.3
pdfminer.six>=20240706
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.1
pydantic>=2.0.0
flask>=2.0.0